+++ app/models/transaction.rb
@@ -1,0 +1,2 @@
+class Transaction < ApplicationRecord
+end
