+++ db/migrate/20221109152739_create_transactions.rb
@@ -1,0 +1,8 @@
+class CreateTransactions < ActiveRecord::Migration[7.0]
+  def change
+    create_table :transactions do |t|
+
+      t.timestamps
+    end
+  end
+end
